cambio a unitedcompany

+++ database/migrations/2017_05_14_151205_entrust_setup_tables.php
@@ -11,49 +11,55 @@ class EntrustSetupTables extends Migration
      */
     public function up()
     {
-        // Create table for storing roles
-        Schema::create('roles', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name')->unique();
-            $table->string('display_name')->nullable();
-            $table->string('description')->nullable();
-            $table->timestamps();
-        });
 
-        // Create table for associating roles to users (Many-to-Many)
-        Schema::create('role_user', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned();
-            $table->integer('role_id')->unsigned();
+        if (!Schema::hasTable('roles') && !Schema::hasTable('role_user') && !Schema::hasTable('permissions')  && !Schema::hasTable('permission_role')  ) {
+            
+            
+            // Create table for storing roles
+            Schema::create('roles', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name')->unique();
+                $table->string('display_name')->nullable();
+                $table->string('description')->nullable();
+                $table->timestamps();
+            });
 
-            $table->foreign('user_id')->references('id')->on('users')
-                ->onUpdate('cascade')->onDelete('cascade');
-            $table->foreign('role_id')->references('id')->on('roles')
-                ->onUpdate('cascade')->onDelete('cascade');
+            // Create table for associating roles to users (Many-to-Many)
+            Schema::create('role_user', function (Blueprint $table) {
+                $table->integer('user_id')->unsigned();
+                $table->integer('role_id')->unsigned();
 
-            $table->primary(['user_id', 'role_id']);
-        });
+                $table->foreign('user_id')->references('id')->on('users')
+                    ->onUpdate('cascade')->onDelete('cascade');
+                $table->foreign('role_id')->references('id')->on('roles')
+                    ->onUpdate('cascade')->onDelete('cascade');
 
-        // Create table for storing permissions
-        Schema::create('permissions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name')->unique();
-            $table->string('display_name')->nullable();
-            $table->string('description')->nullable();
-            $table->timestamps();
-        });
+                $table->primary(['user_id', 'role_id']);
+            });
 
-        // Create table for associating permissions to roles (Many-to-Many)
-        Schema::create('permission_role', function (Blueprint $table) {
-            $table->integer('permission_id')->unsigned();
-            $table->integer('role_id')->unsigned();
+            // Create table for storing permissions
+            Schema::create('permissions', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name')->unique();
+                $table->string('display_name')->nullable();
+                $table->string('description')->nullable();
+                $table->timestamps();
+            });
 
-            $table->foreign('permission_id')->references('id')->on('permissions')
-                ->onUpdate('cascade')->onDelete('cascade');
-            $table->foreign('role_id')->references('id')->on('roles')
-                ->onUpdate('cascade')->onDelete('cascade');
+            // Create table for associating permissions to roles (Many-to-Many)
+            Schema::create('permission_role', function (Blueprint $table) {
+                $table->integer('permission_id')->unsigned();
+                $table->integer('role_id')->unsigned();
 
-            $table->primary(['permission_id', 'role_id']);
-        });
+                $table->foreign('permission_id')->references('id')->on('permissions')
+                    ->onUpdate('cascade')->onDelete('cascade');
+                $table->foreign('role_id')->references('id')->on('roles')
+                    ->onUpdate('cascade')->onDelete('cascade');
+
+                $table->primary(['permission_id', 'role_id']);
+            });
+
+         }
     }
 
     /**
